Add a simple echo feature
This commit allows you to use a simple echo application that prints out
the user input from keyboard until the  key interrupts.

Example
The quick brown fox jumps on the lazy dog.
echo: The quick brown fox jumps on the lazy dog.

Limitation
It does not support a  key (added to TODO list).